+++ spec/rubocop/cops/string_literals_spec.rb
@@ -9,9 +9,7 @@ module Rubocop
 
       it 'registers an offence for double quotes when single quotes suffice' do
         inspect_source(sl, ['s = "abc"'])
-        expect(sl.offences.map(&:message)).to eq(
-          ["Prefer single-quoted strings when you don't need string " +
-           'interpolation or special symbols.'])
+        expect(sl.offences.size).to eq(1)
       end
 
       it 'accepts double quotes when they are needed' do
@@ -22,7 +20,12 @@ module Rubocop
                'e = "#$test"',
                'f = "#@@test"']
         inspect_source(sl, src)
-        expect(sl.offences.map(&:message)).to be_empty
+        expect(sl.offences).to be_empty
+      end
+
+      it 'accepts double quotes at the start of regexp literals' do
+        inspect_source(sl, ['s = /"((?:[^\\"]|\\.)*)"/'])
+        expect(sl.offences).to be_empty
       end
 
       it 'accepts double quotes with some other special symbols' do
@@ -31,7 +34,7 @@ module Rubocop
         # http://www.ruby-doc.org/docs/ProgrammingRuby/html/language.html
         src = ['g = "\xf9"']
         inspect_source(sl, src)
-        expect(sl.offences.map(&:message)).to be_empty
+        expect(sl.offences).to be_empty
       end
 
       it 'can handle double quotes within embedded expression' do
@@ -39,7 +42,7 @@ module Rubocop
         pending do
           src = ['"#{"A"}"']
           inspect_source(sl, src)
-          expect(sl.offences.map(&:message)).to be_empty
+          expect(sl.offences).to be_empty
         end
       end
     end